Added core functionality
Added a readme file so people can see what the program is about and what to do.

Added main.py which contains code using the built-in python librarys and the keyboard library to make stafing. This works by brefly pressing down the a key then d keys at random intervals

Added the Setup.bat file which installs the keyboard library for people that dont know how to do it.

Added the Start.bat file which starts up the python program. Reccomended for people to create a shortcut to start that program to stop having to go to that directory.